perf: drop `execa`
We don't actually use it in the main vite package, so that one is an
easy win.

In the electron package, we use it only to launch a well known path,
i.e. we don't use any features of `execa`. So we should be able to
switch to using the built in `child_process` module directly.